web-scraper-python/requirements.txt
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
requests==2.32.3
soupsieve==2.5
urllib3==2.2.2
python-dotenv==1.0.1
pytest==8.3.2
apscheduler==3.10.4
pandas==2.2.2
pillow==10.4.0
opencv-python-headless==4.10.0.84
